feat: add profile list layout
