+++ src/common/profiles.ts
@@ -1,0 +1,10 @@
+import { IProfileCard } from '@/types/config';
+
+export const ProfileCardList: IProfileCard[] = [
+  {
+    name: 'Alexandre Sanlim',
+    snapshootTime: '2023/11/11 22:36',
+    imageUrl:
+      'https://raw.githubusercontent.com/LHRUN/file-store/main/FireShot%20Capture%20005%20-%20alexandresanlim%20(Alexandre%20Sanlim)%20-%20github.com.png'
+  }
+];
