+++ src/components/componentsCard/components/love/index.tsx
@@ -1,18 +1,16 @@
-'use client';
-
 import { FC, MouseEvent, useMemo } from 'react';
 import classNames from 'classnames';
 import Image from 'next/image';
 import GreyHeart from '@/assets/image/icons/grey-heart.png';
 import SparklingHeart from '@/assets/image/icons/sparkling-heart.png';
-import { ICard } from '@/types/config';
+import { IComponentCard } from '@/types/config';
 import { useList } from '@/context/list';
 import { ACTION_TYPE } from '@/context/list/reducer';
 
 import styles from './index.module.scss';
 
 interface IProps {
-  data: ICard;
+  data: IComponentCard;
   className?: string;
 }
