+++ src/components/componentsCard/index.module.scss
