+++ src/components/profilesList/index.module.scss
@@ -1,0 +1,75 @@
+.list {
+  display: flex;
+  flex-wrap: wrap;
+  margin: 50px 0;
+
+  .item {
+    position: relative;
+    width: 45%;
+    padding: 15px;
+    overflow: hidden;
+    cursor: pointer;
+    border-radius: 20px;
+
+    .bg {
+      position: absolute;
+      top: 0;
+      right: 0;
+      bottom: 0;
+      left: 0;
+      z-index: 1;
+      width: 100%;
+      height: 100%;
+      background-color: $primary-color;
+    }
+
+    &:hover {
+      .bg {
+        opacity: 0.8;
+      }
+    }
+
+    &:not(:nth-child(2n)) {
+      margin-right: 5%;
+    }
+
+    &:not(:nth-child(-n+2)) {
+      margin-top: 30px;
+    }
+
+    .header {
+      position: relative;
+      z-index: 2;
+      .name {
+        font-size: 26px;
+        line-height: 28px;
+        color: $light-color;
+      }
+      .time {
+        font-size: 15px;
+        line-height: 22px;
+        color: $tertiary-color;
+
+        .timeText {
+          display: inline-block;
+          margin-left: 5px;
+        }
+      }
+    }
+
+    .imageContainer {
+      position: relative;
+      z-index: 2;
+      width: 100%;
+      height: 300px;
+      margin-top: 5px;
+      overflow: hidden;
+      border-radius: 10px;
+
+      .image {
+        width: 100%;
+        height: 100%;
+      }
+    }
+  }
+}
