+++ src/components/tabs/index.module.scss
@@ -1,0 +1,71 @@
+$max-tab-height: 42px;
+$tab-height: 32px;
+
+.tab-list {
+	position: relative;
+  z-index: 2;
+	display: flex;
+  flex-direction: row-reverse;
+  align-items: flex-end;
+  height: $max-tab-height;
+  margin-top: 30px;
+  margin-left: auto;
+  background-color: #000000;
+
+	.tab-item {
+		position: relative;
+		font-size: 28px;
+		font-weight: 600;
+    line-height: $max-tab-height;
+		color: $light-color;
+		cursor: pointer;
+    border-top-left-radius: 15px;
+    border-top-right-radius: 15px;
+    transition: all 0.2s;
+
+    .firstChat {
+      position: relative;
+      top: 2px;
+      display: inline-block;
+      margin-right: 2px;
+      font-size: 40px;
+      transform: rotate(0deg);
+    }
+
+    &.Profiles {
+      background-color: $secondary-color;
+    }
+    &.Components {
+      background-color: $primary-color;
+    }
+
+    &:not(:first-child) {
+      transform: translateX(50%);
+    }
+	}
+
+  .tab-selected {
+    z-index: 2;
+    height: $max-tab-height;
+    padding: 0 30px;
+    opacity: 1;
+
+    .firstChat {
+      top: -3px;
+      margin-right: 6px;
+      font-size: 60px;
+      transform: rotate(28deg);
+    }
+  }
+
+  .tab-no-selected {
+    z-index: 1;
+    height: $tab-height;
+    padding: 0 15px;
+    opacity: 0.6;
+
+    &:hover {
+      opacity: 0.9;
+    }
+  }
+}
